Fix undefined behaviour from left shift.
Found by Coverity Scan.

+++ libheif/heif_avif.cc
@@ -87,7 +87,7 @@ static uint64_t leb128(BitReader& reader)
 {
   uint64_t val = 0;
   for (int i = 0; i < 8; i++) {
-    int v = reader.get_bits(8);
+    int64_t v = reader.get_bits(8);
     val |= (v & 0x7F) << (i * 7);
     if (!(v & 0x80)) {
       break;
